profile page – only shows chirps by user

+++ main/views.py
@@ -15,7 +15,8 @@ def main_view(request):
     # if post request, insert into DB
     if request.method == 'POST' and request.POST['body'] != "":
         chirp = Chirp.objects.create(
-            body = request.POST['body']
+            body = request.POST['body'],
+            author = request.user
         )
 
     chirps = Chirp.objects.all()
@@ -23,16 +24,16 @@ def main_view(request):
 
 # delete a chirp routing
 def delete_view(request):
-    # chirp = Chirps.objects.get(id=request.POST[], request=)
     chirp = Chirp.objects.get(id=request.GET['id'])
-    chirp.delete()
+    if chirp.author == request.user:
+        chirp.delete()
     return redirect('/')
-
 
 # accounts login/signup page
 def accounts_view(request):
     return render(request, 'accounts.html', {})
 
+# login a user
 def login_view(request):
     username, password = request.POST['username'], request.POST['password']
     user = authenticate(username = username, password = password)
@@ -43,6 +44,7 @@ def login_view(request):
     else:
         return redirect('/accounts?error=True')
 
+# register a user
 def signup_view(request):
     user = User.objects.create_user(
         username = request.POST['username'],
@@ -53,6 +55,12 @@ def signup_view(request):
     login(request, user)
     return redirect('/')
 
+# logout a user
 def logout_view(request):
     logout(request)
     return redirect('/accounts/')
+
+# profile page
+def profile_view(request):
+    chirps = Chirp.objects.filter(author=request.user)
+    return render(request, 'main.html', {'chirps': chirps})
